add abort in delete_question method

+++ starter/backend/flaskr/__init__.py
@@ -97,13 +97,19 @@ def create_app(test_config=None):
   '''
   @app.route('/questions/<int:question_id>', methods=['DELETE'])
   def delete_question(question_id):
-    question = Question.query.filter_by(Question.id==question_id).one_or_none()
-    question.delete()
-    
-    return jsonify({
-      'success' : True,
-      'deleted_question_id' : question_id
-    })
+    try:
+      question = Question.query.filter_by(Question.id==question_id).one_or_none()
+      if question is None:
+        abort(404)
+      
+      question.delete()
+      
+      return jsonify({
+        'success' : True,
+        'deleted_question' : question_id
+      })
+    except:
+      abort(422)
 
   '''
   @TODO:
